requirements.txt
streamlit
openai
python-dotenv

streamlit_webrtc
git+https://github.com/openai/whisper.git
